hookUpdate: creating a context
