+++ src/components/shared/interfaces.tsx
@@ -1,0 +1,9 @@
+export interface Cell {
+  x: number;
+  y: number;
+  num: number;
+}
+
+export interface Grid {
+  grid: Cell[];
+}
